refactor: rename index.ts of core fhir's org/patient to conversion.ts
Ref. metriport/metriport-internal#2215
