+++ packages/api/src/__tests__/e2e/mapi/parts/patient.ts
@@ -3,7 +3,7 @@ import { Address, Patient as FhirPatient } from "@medplum/fhirtypes";
 import { USState } from "@metriport/shared";
 import { PatientCreate, PatientDTO } from "@metriport/api-sdk";
 import { Patient } from "@metriport/core/domain/patient";
-import { mapMetriportGenderToFhirGender } from "@metriport/core/external/fhir/patient/index";
+import { mapMetriportGenderToFhirGender } from "@metriport/core/external/fhir/patient/conversion";
 import { PatientWithId } from "@metriport/core/external/fhir/__tests__/patient";
 
 export const createPatient: PatientCreate = {
