+++ packages/api/src/command/medical/patient/handle-data-contributions.ts
@@ -1,8 +1,8 @@
 import { Bundle, Resource } from "@medplum/fhirtypes";
 import { createUploadFilePath, FHIR_BUNDLE_SUFFIX } from "@metriport/core/domain/document/upload";
 import { Patient } from "@metriport/core/domain/patient";
-import { toFHIR as toFhirOrganization } from "@metriport/core/external/fhir/organization/index";
-import { toFHIR as toFhirPatient } from "@metriport/core/external/fhir/patient/index";
+import { toFHIR as toFhirOrganization } from "@metriport/core/external/fhir/organization/conversion";
+import { toFHIR as toFhirPatient } from "@metriport/core/external/fhir/patient/conversion";
 import { uploadCdaDocuments, uploadFhirBundleToS3 } from "@metriport/core/fhir-to-cda/upload";
 import { out } from "@metriport/core/util/log";
 import { uuidv7 } from "@metriport/core/util/uuid-v7";
