+++ packages/api/src/external/fhir/document/draft-update-document-reference.ts
@@ -7,8 +7,8 @@ import {
   Resource,
 } from "@medplum/fhirtypes";
 import { S3Utils } from "@metriport/core/external/aws/s3";
-import { appendIdentifierOID, toFHIR } from "@metriport/core/external/fhir/organization/index";
-import { toFHIRSubject } from "@metriport/core/external/fhir/patient/index";
+import { appendIdentifierOID, toFHIR } from "@metriport/core/external/fhir/organization/conversion";
+import { toFHIRSubject } from "@metriport/core/external/fhir/patient/conversion";
 import { metriportDataSourceExtension } from "@metriport/core/external/fhir/shared/extensions/metriport";
 import { IETF_URI } from "@metriport/core/external/fhir/shared/namespaces";
 import BadRequestError from "@metriport/core/util/error/bad-request";
