+++ packages/api/src/external/fhir/document/index.ts
@@ -24,17 +24,18 @@ import {
   HumanName as CWHumanName,
 } from "@metriport/commonwell-sdk";
 import { Gender } from "@metriport/commonwell-sdk/src/models/demographics";
-import { metriportDataSourceExtension } from "@metriport/core/external/fhir/shared/extensions/metriport";
+import { joinName, Patient, splitName } from "@metriport/core/domain/patient";
 import { cwExtension } from "@metriport/core/external/commonwell/extension";
+import { toFHIRSubject } from "@metriport/core/external/fhir/patient/conversion";
+import { metriportDataSourceExtension } from "@metriport/core/external/fhir/shared/extensions/metriport";
 import dayjs from "dayjs";
 import isToday from "dayjs/plugin/isToday";
 import { sortBy, uniqBy } from "lodash";
-import { joinName, Patient, splitName } from "@metriport/core/domain/patient";
 import MetriportError from "../../../errors/metriport-error";
 import { capture } from "../../../shared/notifications";
 import { Util } from "../../../shared/util";
 import { CWDocumentWithMetriportData } from "../../commonwell/document/shared";
-import { toFHIRSubject } from "@metriport/core/external/fhir/patient/index";
+
 dayjs.extend(isToday);
 
 export const MAX_FHIR_DOC_ID_LENGTH = 64;
