+++ packages/core/src/command/consolidated/get-snapshot-local.ts
@@ -5,8 +5,8 @@ import axios from "axios";
 import { analytics, EventTypes } from "../../external/analytics/posthog";
 import { isConsolidatedFromS3Enabled } from "../../external/aws/app-config";
 import { getConsolidatedFhirBundle as getConsolidatedFromFhirServer } from "../../external/fhir/consolidated/consolidated";
-import { toFHIR as organizationToFhir } from "../../external/fhir/organization/index";
-import { toFHIR as patientToFhir } from "../../external/fhir/patient/index";
+import { toFHIR as organizationToFhir } from "../../external/fhir/organization/conversion";
+import { toFHIR as patientToFhir } from "../../external/fhir/patient/conversion";
 import { buildBundleEntry } from "../../external/fhir/shared/bundle";
 import { deduplicateFhir } from "../../fhir-deduplication/deduplicate-fhir";
 import { out } from "../../util";
