+++ packages/core/src/external/carequality/ihe-gateway-v2/patient.ts
@@ -1,4 +1,4 @@
-import { PatientIdAndData, mapPatientDataToResource } from "../../fhir/patient/index";
+import { PatientIdAndData, mapPatientDataToResource } from "../../fhir/patient/conversion";
 import { PatientResource } from "@metriport/ihe-gateway-sdk";
 
 export function toIheGatewayPatientResource(patient: PatientIdAndData): PatientResource {
