+++ packages/shared/src/common/error.ts
@@ -1,7 +1,7 @@
 export type ErrorToStringOptions = { detailed: boolean };
 
 /**
- * @deprecated User @metriport/shared/error/shared instead
+ * @deprecated Use `@metriport/shared/error/shared` instead
  */
 export function errorToString(
   err: unknown,
@@ -14,14 +14,14 @@ export function errorToString(
 }
 
 /**
- * @deprecated User @metriport/shared/error/shared instead
+ * @deprecated Use `@metriport/shared/error/shared` instead
  */
 export function genericErrorToString(err: unknown): string {
   return (err as any)["message"] ?? String(err); // eslint-disable-line @typescript-eslint/no-explicit-any
 }
 
 /**
- * @deprecated User @metriport/shared/error/shared instead
+ * @deprecated Use `@metriport/shared/error/shared` instead
  */
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export function detailedErrorToString(error: any): string {
